Move dtls cid tests to tests/api/dtls.c
