+++ tests/api/include.am
@@ -13,6 +13,7 @@ tests_unit_test_SOURCES += tests/api/test_sm3.c
 tests_unit_test_SOURCES += tests/api/test_ripemd.c
 tests_unit_test_SOURCES += tests/api/test_hash.c
 tests_unit_test_SOURCES += tests/api/test_ascon.c
+tests_unit_test_SOURCES += tests/api/test_dtls.c
 endif
 EXTRA_DIST += tests/api/api.h
 EXTRA_DIST += tests/api/test_md5.h
@@ -27,4 +28,5 @@ EXTRA_DIST += tests/api/test_hash.h
 EXTRA_DIST += tests/api/test_ascon.h
 EXTRA_DIST += tests/api/test_ascon.h
 EXTRA_DIST += tests/api/test_ascon_kats.h
+EXTRA_DIST += tests/api/test_dtls.h
